Improve testing framework
This commit introduces a number of improvements to the run_test_script
program, most of them meant to bring it closer into line with the Python
doctest module. It now compares for string equality rather than value
equality, so if the calculated value is 15, then (* 3 5) will no longer
suffice. It also prints more helpful error messages and doesn't abort
after the first error.