Add more tests to tests.c and fix some mixed arithmetic bugs
I know I've vacillated on this a lot, but I think that having most of
the tests in src/tests.c is a good idea, because it keeps the
complexity of the test apparatus to an absolute minimum. The
disadvantage is that tests needs to be recompiled with the interpreter
every time, but I think that it should run faster and make up for the
difference.